requirements.txt
aiobotocore==2.4.1
aiohttp
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.23
botocore==1.27.59
charset-normalizer==2.1.1
et-xmlfile==1.1.0
frozenlist==1.3.3
fsspec==2022.11.0
greenlet==2.0.1
idna==3.4
jmespath==1.0.1
multidict==6.0.3
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
s3fs==2022.11.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.2
pytest==7.1.2
boto3
pandas
python-dotenv
sqlalchemy
psycopg2-binary
openpyxl
s3fs
